Implement notification functionality
I implemented two notifications routes one to mark one notification as read and another to mark all as read. I also modified sending a message,
creating a group, adding and removing members from a group and appointing group admins so that the affected user receive a notification in regard
to what just happened.

Read field served no purpose in the notification model so I removed it and instead of modifying it to true when read, I delete it all together.